chore: rename to data-view-beta
