+++ packages/raystack/components/data-view-beta/components/display-properties.tsx
